Refactor camera limiter code
