+++ src/components/molecules/Visualizer/Engine/Cesium/hooks.ts
@@ -1,16 +1,4 @@
-import {
-  Color,
-  Entity,
-  Ion,
-  Cesium3DTileFeature,
-  Cartesian3,
-  Cartographic,
-  Camera as CesiumCamera,
-  Math,
-  EllipsoidGeodesic,
-  Rectangle,
-  PolylineDashMaterialProperty,
-} from "cesium";
+import { Color, Entity, Ion, Cesium3DTileFeature, Cartesian3 } from "cesium";
 import type { Viewer as CesiumViewer, ImageryProvider, TerrainProvider } from "cesium";
 import CesiumDnD, { Context } from "cesium-dnd";
 import { isEqual, throttle } from "lodash-es";
@@ -23,6 +11,7 @@ import { Camera, LatLng } from "@reearth/util/value";
 
 import type { SelectLayerOptions, Ref as EngineRef, SceneProperty } from "..";
 
+import { useCameraLimiter } from "./cameraLimiter";
 import { getCamera, isDraggable, isSelectable, layerIdField } from "./common";
 import imagery from "./imagery";
 import terrain from "./terrain";
@@ -136,10 +125,10 @@ export default ({
       if (initialCameraFlight.current) return;
       initialCameraFlight.current = true;
       if (
-        property?.cameraLimiter?.cameraLimitterEnabled &&
-        property?.cameraLimiter?.cameraLimitterTargetArea
+        property?.cameraLimiter?.cameraLimiterEnabled &&
+        property?.cameraLimiter?.cameraLimiterTargetArea
       ) {
-        engineAPI.flyTo(property?.cameraLimiter?.cameraLimitterTargetArea, { duration: 0 });
+        engineAPI.flyTo(property?.cameraLimiter?.cameraLimiterTargetArea, { duration: 0 });
       } else if (property?.default?.camera) {
         engineAPI.flyTo(property.default.camera, { duration: 0 });
       }
@@ -152,7 +141,7 @@ export default ({
       engineAPI,
       onCameraChange,
       property?.default?.camera,
-      property?.cameraLimiter?.cameraLimitterEnabled,
+      property?.cameraLimiter?.cameraLimiterEnabled,
     ],
     (prevDeps, nextDeps) =>
       prevDeps[0] === nextDeps[0] &&
@@ -193,179 +182,6 @@ export default ({
       emittedCamera.current = undefined;
     }
   }, [camera, engineAPI]);
-
-  const geodsic = useMemo(():
-    | undefined
-    | { geodesicVertical: EllipsoidGeodesic; geodesicHorizontal: EllipsoidGeodesic } => {
-    const viewer = cesium.current?.cesiumElement;
-    if (
-      !viewer ||
-      !viewer.scene ||
-      viewer.isDestroyed() ||
-      !property?.cameraLimiter ||
-      !property?.cameraLimiter.cameraLimitterTargetArea
-    )
-      return undefined;
-    const ellipsoid = viewer.scene.globe.ellipsoid;
-
-    const centerPoint = Cartesian3.fromDegrees(
-      property.cameraLimiter.cameraLimitterTargetArea.lng,
-      property.cameraLimiter.cameraLimitterTargetArea.lat,
-      0,
-    );
-
-    const cartographicCenterPoint = Cartographic.fromCartesian(centerPoint);
-    const normal = ellipsoid.geodeticSurfaceNormal(centerPoint);
-    const east = Cartesian3.normalize(
-      Cartesian3.cross(Cartesian3.UNIT_Z, normal, new Cartesian3()),
-      new Cartesian3(),
-    );
-    const north = Cartesian3.normalize(
-      Cartesian3.cross(normal, east, new Cartesian3()),
-      new Cartesian3(),
-    );
-
-    const geodesicVertical = new EllipsoidGeodesic(
-      cartographicCenterPoint,
-      Cartographic.fromCartesian(north),
-      ellipsoid,
-    );
-    const geodesicHorizontal = new EllipsoidGeodesic(
-      cartographicCenterPoint,
-      Cartographic.fromCartesian(east),
-      ellipsoid,
-    );
-    return { geodesicVertical, geodesicHorizontal };
-  }, [property?.cameraLimiter]);
-
-  // calculate inner limiter dimensions
-  const targetWidth = 1000000;
-  const targetLength = 1000000;
-  const limiterDimensions = useMemo(():
-    | undefined
-    | {
-        cartographicDimensions: {
-          rightDemention: Cartographic;
-          leftDemention: Cartographic;
-          topDemention: Cartographic;
-          bottomDemention: Cartographic;
-        };
-        cartesianArray: Cartesian3[];
-      } => {
-    const viewer = cesium.current?.cesiumElement;
-    if (
-      !viewer ||
-      viewer.isDestroyed() ||
-      !property?.cameraLimiter ||
-      !property?.cameraLimiter.cameraLimitterTargetArea ||
-      !geodsic
-    )
-      return undefined;
-    const {
-      cameraLimitterTargetWidth: width = targetWidth,
-      cameraLimitterTargetLength: length = targetLength,
-    } = property?.cameraLimiter ?? {};
-
-    const { cartesianArray, cartographicDimensions } = calcBoundaryBox(
-      geodsic,
-      length / 2,
-      width / 2,
-    );
-
-    return {
-      cartographicDimensions,
-      cartesianArray,
-    };
-  }, [property?.cameraLimiter, geodsic]);
-
-  // calculate maximum camera view (outer boundaries)
-  const [cameraViewOuterBoundaries, setCameraViewOuterBoundaries] = useState<
-    Cartesian3[] | undefined
-  >();
-
-  useEffect(() => {
-    const viewer = cesium.current?.cesiumElement;
-    if (
-      !viewer ||
-      viewer.isDestroyed() ||
-      !property?.cameraLimiter?.cameraLimitterTargetArea ||
-      !geodsic
-    )
-      return;
-
-    const camera = new CesiumCamera(viewer.scene);
-    camera.setView({
-      destination: Cartesian3.fromDegrees(
-        property.cameraLimiter.cameraLimitterTargetArea.lng,
-        property.cameraLimiter.cameraLimitterTargetArea.lat,
-        property.cameraLimiter.cameraLimitterTargetArea.height,
-      ),
-      orientation: {
-        heading: property?.cameraLimiter?.cameraLimitterTargetArea.heading,
-        pitch: property?.cameraLimiter?.cameraLimitterTargetArea.pitch,
-        roll: property?.cameraLimiter?.cameraLimitterTargetArea.roll,
-        up: camera.up,
-      },
-    });
-
-    const computedViewRectangle = camera.computeViewRectangle();
-    if (!computedViewRectangle) return;
-    const rectangleHalfWidth = Rectangle.computeWidth(computedViewRectangle) * Math.PI * 1000000;
-    const rectangleHalfHeight = Rectangle.computeHeight(computedViewRectangle) * Math.PI * 1000000;
-
-    const {
-      cameraLimitterTargetWidth: width = targetWidth,
-      cameraLimitterTargetLength: length = targetLength,
-    } = property?.cameraLimiter ?? {};
-
-    const { cartesianArray } = calcBoundaryBox(
-      geodsic,
-      length / 2 + rectangleHalfHeight,
-      width / 2 + rectangleHalfWidth,
-    );
-
-    setCameraViewOuterBoundaries(cartesianArray);
-  }, [property?.cameraLimiter, geodsic, camera]);
-
-  // Manage camera limiter conditions
-  useEffect(() => {
-    const camera = getCamera(cesium?.current?.cesiumElement);
-    const viewer = cesium?.current?.cesiumElement;
-    if (
-      !viewer ||
-      viewer.isDestroyed() ||
-      !property?.cameraLimiter?.cameraLimitterEnabled ||
-      !limiterDimensions
-    )
-      return;
-    if (camera) {
-      const cameraPosition = Cartographic.fromDegrees(camera?.lng, camera?.lat, camera?.height);
-
-      const destination = new Cartographic(
-        Math.clamp(
-          cameraPosition.longitude,
-          limiterDimensions.cartographicDimensions.leftDemention.longitude,
-          limiterDimensions.cartographicDimensions.rightDemention.longitude,
-        ),
-        Math.clamp(
-          cameraPosition.latitude,
-          limiterDimensions.cartographicDimensions.bottomDemention.latitude,
-          limiterDimensions.cartographicDimensions.topDemention.latitude,
-        ),
-        cameraPosition.height,
-      );
-
-      viewer.camera.setView({
-        destination: Cartographic.toCartesian(destination),
-        orientation: {
-          heading: viewer.camera.heading,
-          pitch: viewer.camera.pitch,
-          roll: viewer.camera.roll,
-          up: viewer.camera.up,
-        },
-      });
-    }
-  }, [camera, property?.cameraLimiter, limiterDimensions]);
 
   // manage layer selection
   useEffect(() => {
@@ -466,6 +282,8 @@ export default ({
       cesiumDnD.current?.disable();
     };
   }, [handleLayerDrag, handleLayerDrop, isLayerDraggable]);
+  const { limiterDimensions, cameraViewOuterBoundaries, cameraViewBoundariesMaterial } =
+    useCameraLimiter(cesium, camera, property?.cameraLimiter);
 
   return {
     terrainProvider,
@@ -483,45 +301,6 @@ export default ({
   };
 };
 
-function calcBoundaryBox(
-  geodsic: { geodesicVertical: EllipsoidGeodesic; geodesicHorizontal: EllipsoidGeodesic },
-  halfLength: number,
-  halfWidth: number,
-): {
-  cartographicDimensions: {
-    rightDemention: Cartographic;
-    leftDemention: Cartographic;
-    topDemention: Cartographic;
-    bottomDemention: Cartographic;
-  };
-  cartesianArray: Cartesian3[];
-} {
-  const topDemention = geodsic.geodesicVertical.interpolateUsingSurfaceDistance(halfLength);
-  const bottomDemention = geodsic.geodesicVertical.interpolateUsingSurfaceDistance(-halfLength);
-  const rightDemention = geodsic.geodesicHorizontal.interpolateUsingSurfaceDistance(halfWidth);
-  const leftDemention = geodsic.geodesicHorizontal.interpolateUsingSurfaceDistance(-halfWidth);
-
-  const rightTop = new Cartographic(rightDemention.longitude, topDemention.latitude, 0);
-  const leftTop = new Cartographic(leftDemention.longitude, topDemention.latitude, 0);
-  const rightBottom = new Cartographic(rightDemention.longitude, bottomDemention.latitude, 0);
-  const leftBottom = new Cartographic(leftDemention.longitude, bottomDemention.latitude, 0);
-  return {
-    cartographicDimensions: {
-      rightDemention,
-      leftDemention,
-      topDemention,
-      bottomDemention,
-    },
-    cartesianArray: [
-      Cartographic.toCartesian(rightTop),
-      Cartographic.toCartesian(leftTop),
-      Cartographic.toCartesian(leftBottom),
-      Cartographic.toCartesian(rightBottom),
-      Cartographic.toCartesian(rightTop),
-    ],
-  };
-}
-
 function tileProperties(t: Cesium3DTileFeature): { key: string; value: any }[] {
   return t
     .getPropertyNames()
@@ -530,10 +309,6 @@ function tileProperties(t: Cesium3DTileFeature): { key: string; value: any }[] {
       [],
     );
 }
-
-const cameraViewBoundariesMaterial = new PolylineDashMaterialProperty({
-  color: Color.RED,
-});
 
 function findEntity(viewer: CesiumViewer, layerId: string | undefined): Entity | undefined {
   let entity: Entity | undefined;
